src/main/java/com/qianhong/calculator/CalculatorService.java
package com.qianhong.calculator;

import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.QueryParam;
import javax.ws.rs.core.MediaType;
import java.util.Date;

// Define a RESTful web service for calculator operations.
@Path("/calculator")
public class CalculatorService {

    // Handle HTTP GET requests to "/calculator/ping" and return a plain text response.
    @GET
    @Path("ping")
    @Produces(MediaType.TEXT_PLAIN)
    public String ping() {
        return "Welcome to Java Maven Calculator Web App!!!\n" + new Date().toString();
    }

    // Handle HTTP GET requests to "/calculator/add" and return a JSON response with the result of addition.
    @GET
    @Path("add")
    @Produces(MediaType.APPLICATION_JSON)
    public CalculatorResponse Add(@QueryParam("x") int x, @QueryParam("y") int y) {
        return new CalculatorResponse(x, y, x + y);
    }

    // Handle HTTP GET requests to "/calculator/sub" and return a JSON response with the result of subtraction.
    @GET
    @Path("sub")
    @Produces(MediaType.APPLICATION_JSON)
    public CalculatorResponse Sub(@QueryParam("x") int x, @QueryParam("y") int y) {
        return new CalculatorResponse(x, y, x - y);
    }

    // Handle HTTP GET requests to "/calculator/mul" and return a JSON response with the result of multiplication.
    @GET
    @Path("mul")
    @Produces(MediaType.APPLICATION_JSON)
    public CalculatorResponse Mul(@QueryParam("x") int x, @QueryParam("y") int y) {
        return new CalculatorResponse(x, y, x * y);
    }

    // Handle HTTP GET requests to "/calculator/div" and return a JSON response with the result of division.
    @GET
    @Path("div")
    @Produces(MediaType.APPLICATION_JSON)
    public CalculatorResponse Div(@QueryParam("x") int x, @QueryParam("y") int y) {
        return new CalculatorResponse(x, y, x / y);
    }
}
